Add item update handler

+++ src/main/java/ca/yorku/item/DeleteItemHandler.java
@@ -1,5 +1,6 @@
 package ca.yorku.item;
 
+import ca.yorku.dal.Item;
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.serverless.ApiGatewayResponse;
@@ -10,11 +11,22 @@ public class DeleteItemHandler implements RequestHandler<Map<String, Object>, Ap
 
     @Override
     public ApiGatewayResponse handleRequest(Map<String, Object> input, Context context) {
-        ApiGatewayResponse res =  ApiGatewayResponse.builder()
-                .setStatusCode(200)
-                .setObjectBody("DeleteItem")
-                .build();
-        return res;
+        Map<String,String> pathParameters =  (Map<String,String>)input.get("pathParameters");
+        String itemId = pathParameters.get("itemId");
+
+        boolean deleted = new Item().delete(itemId);
+        if(deleted) {
+            return ApiGatewayResponse.builder()
+                    .setStatusCode(200)
+                    .setObjectBody("DeleteItem")
+                    .build();
+        }
+        else{
+            return ApiGatewayResponse.builder()
+                    .setStatusCode(404)
+                    .setRawBody("Item not found")
+                    .build();
+        }
 
     }
 }
